Rewrite code, add more networking shenanigans
Not sure if clients can see the server authorization token or not; if they can then that's a big security issue that would need to be resolved. I can't think of a way the client could access the token though, so I think it's fine. Probably.
Fixed the known issue from the first commit; clients that are not the main client can now directly converse with AI via the ping method.
Updated default admin OP level to 3 as I found it more suitable to the level of power that it grants you over the AI.